% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agc.R
\name{mod_agc_server}
\alias{mod_agc_server}
\title{Módulo de perguntas sobre AGC
Informações sobre AGC}
\usage{
mod_agc_server(id, input_mod_processo)
}
\arguments{
\item{id}{id}

\item{input_mod_processo}{input_mod_processo}
}
\description{
Módulo de perguntas sobre AGC
Informações sobre AGC
}
