% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_aj.R
\name{mod_aj_server}
\alias{mod_aj_server}
\title{Módulo de perguntas sobre o AJ
Informações sobre o administrador judicial}
\usage{
mod_aj_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre o AJ
Informações sobre o administrador judicial
}
