% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_credores.R
\name{mod_credores_server}
\alias{mod_credores_server}
\title{Módulo de perguntas sobre credores
Informações sobre credores}
\usage{
mod_credores_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre credores
Informações sobre credores
}
