% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_credores.R
\name{mod_credores_ui}
\alias{mod_credores_ui}
\title{Módulo de perguntas sobre credores
Informações sobre credores}
\usage{
mod_credores_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre credores
Informações sobre credores
}
