% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_deferimento.R
\name{mod_deferimento_server}
\alias{mod_deferimento_server}
\title{Módulo de perguntas sobre deferimento
Informações sobre deferimento}
\usage{
mod_deferimento_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre deferimento
Informações sobre deferimento
}
