% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_deferimento.R
\name{mod_deferimento_ui}
\alias{mod_deferimento_ui}
\title{Módulo de perguntas sobre deferimento
Informações sobre deferimento}
\usage{
mod_deferimento_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre deferimento
Informações sobre deferimento
}
