% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emenda.R
\name{mod_emenda_server}
\alias{mod_emenda_server}
\title{Módulo de perguntas sobre emenda
Informações sobre emenda}
\usage{
mod_emenda_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre emenda
Informações sobre emenda
}
