% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_info.R
\name{mod_info_server}
\alias{mod_info_server}
\title{Módulo de perguntas sobre processo
Escopo e informações sobre o processo}
\usage{
mod_info_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre processo
Escopo e informações sobre o processo
}
