% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_litisconsorcio.R
\name{mod_litisconsorcio_ui}
\alias{mod_litisconsorcio_ui}
\title{Módulo de perguntas sobre litisconsorcio
Informações sobre litisconsorcio}
\usage{
mod_litisconsorcio_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre litisconsorcio
Informações sobre litisconsorcio
}
