% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_pericia.R
\name{mod_pericia_ui}
\alias{mod_pericia_ui}
\title{Módulo de perguntas sobre perícia
Informações sobre perícia}
\usage{
mod_pericia_ui(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre perícia
Informações sobre perícia
}
