% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_stay.R
\name{mod_stay_server}
\alias{mod_stay_server}
\title{Módulo de perguntas sobre stay period
Informações sobre stay period}
\usage{
mod_stay_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre stay period
Informações sobre stay period
}
