% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_agc.R
\name{mod_agc_server}
\alias{mod_agc_server}
\title{Módulo de perguntas sobre AGC
Informações sobre AGC}
\usage{
mod_agc_server(id)
}
\arguments{
\item{id}{id}
}
\description{
Módulo de perguntas sobre AGC
Informações sobre AGC
}
